+++ Changelog.md
@@ -68,7 +68,7 @@ This cursed update adds, well, curses. Curses are another type of effect and gen
 * Added requires_attunement to effects. If the item does not already have an attunement requirement, these effects will not generate.
 * Changed some more configurations to be more readable.
 * Removed requireaffix option for items. It wasn't being used in code, and only one effect (On scrolls) was attempting to use it. With tweaks to scroll code, this shouldn't be needed, and can be accomplished in other ways if it is.
-## 2021.X.X: System Folders
+## 2021.4.7: System Folders
 This update is a simple one, as far as the outside goes. It changes the folders from local folders to a system folder.
 * added startup.cfg, to a (local) configs folder. This is checked on startup and determines if the program is in folder mode, log settings, and logging mode.
 * If local_mode in startup.cfg is set to TRUE, the program will run with local files.